Cast integer numerator to double and make integer denominator a double to make sure double division is performed instead of integer division.  (Yeah, just one would be enough, but makes it clearer.)

+++ SanityChecker/src/uk/ac/uea/socat/sanitychecker/sanitychecks/SpeedSanityCheck.java
@@ -96,6 +96,6 @@ public class SpeedSanityCheck extends SanityCheck {
 	
 	private double calcHourDiff(DateTime time1, DateTime time2) {
 		Seconds diffSeconds = Seconds.secondsBetween(time1, time2);
-		return diffSeconds.getSeconds() / 3600;
+		return ((double) diffSeconds.getSeconds()) / 3600.0;
 	}
 }
